python-ecosys/requests: Make response headers dict lowercase.
The HTTP response headers are case-insensitive.  But the old implementation
put them in a dictionary keyed by case-sensitive header name.  The only way
to correctly get a header would have been to iterate through the dictionary
looking for it, which is slow and a lot of code and I doubt anyone bothered,
I expect that existing code probably just did a case-sensitive match.  That's
fragile, server changes could break it.

To fix that, lowercase all the header names before putting them in the
dictionary.  That way, clients can just access the dictionary with the
lowercase header name and it will do the right thing.

This is backward compatible with correctly-written existing code.  But it's
not compatible with existing code that was buggy anyway and did a
case-sensitive match.